Move read functions to EventWorkerStore
